Test MainViewController instantiation
