--- ExamEGERussianTests/ExamEGERussianTests.swift
@@ -1,22 +1,0 @@
-//
-//  ExamEGERussianTests.swift
-//  ExamEGERussianTests
-//
-//  Created by jetbrains on 03/07/2018.
-//  Copyright © 2018 Alex Karpov. All rights reserved.
-//
-
-import XCTest
-@testable import ExamEGERussian
-
-class ExamEGERussianTests: XCTestCase {
-    
-    override func setUp() {
-        super.setUp()
-    }
-    
-    override func tearDown() {
-        super.tearDown()
-    }
-    
-}
